+++ lib/rockette/commands/deploy.rb
@@ -1,51 +1,53 @@
 # frozen_string_literal: true
 
-require_relative '../command'
+require_relative "../command"
 
 module Rockette
   module Commands
+    # Push export file and import into APEX
     class Deploy < Rockette::Command
+      include TextHelper
+
       def initialize(options)
+        super()
         @options = options
+        @filey = "f#{@options[:app_id]}.sql"
       end
 
-       # Import application export into APEX instance
-      def import_app(url, body)
-        response = RestClient.post url, body
-      rescue StandardError => e
-        puts "Unable to import application because #{e.message}" # puts error
-      end
-
-      # Push attachment to api endpoint and show code
-      def push_blob(filey, headers, url)
-        response = RestClient.post url, filey, headers
+      def importer
+        body = CONF["deploy_body"]
+        body["app_id_src"] = @options[:app_id]
+        body["app_id_tgt"] = @options[:app_id]
+        # body["app_id_tgt"] = "0" #test app copy
+        url = "#{@options[:url]}deploy/app"
+        response = Rester.new(meth: "Post", params: body, url: url).rest_try
+        bail unless response
+        abort padder("Error. Got back response code: #{response.code}") unless (200..201).include? response.code
         response
-      rescue StandardError => e
-        puts "Unable to push file because #{e.message}" # puts error
+      end
+
+      def pusher
+        push_hdrs = CONF["push_hdrs"]
+        push_hdrs["file_name"] = @filey
+        push_url = "#{@options[:url]}data_loader/blob"
+        # Push the chosen export file to the target system
+        filey = File.join(EXPORT_DIR, @filey)
+        response = Rester.new(headers: push_hdrs, meth: "Post", params: File.open(filey), url: push_url).rest_try
+        bail unless response
+        abort padder("Error. Got back response code: #{response.code}") unless (200..201).include? response.code
+        response
       end
 
       def execute(input: $stdin, output: $stdout)
+        check_input(input)
         # Create and download export
-        output.puts "OK, deploying export file #{@options[:file]}"
-        filey = "f#{@options[:app_id]}.sql"
-        # First push the chosen export file to the target system.
-        push_hdrs = CONF["push_hdrs"]
-        push_hdrs["file_name"] = filey
-        push_url = @options[:url] + 'data_loader/blob'
-        pusher = push_blob(File.open(filey), push_hdrs, push_url)
+        output.puts padder("OK, deploying export file #{@filey}")
+        pusher
+        output.puts padder("Pushed #{@filey} to instance and attempting import now...")
         # If push was successful, request application import
-        if pusher.code == 200 || pusher.code == 201
-          sleep 1 # replace with check for file on server
-          puts "Pushed #{filey} and attempting import now..."
-          body = CONF["deploy_body"]
-          body["app_id_src"] = @options[:app_id]
-          body["app_id_tgt"] = @options[:app_id]
-          body["app_id_tgt"] = "0" #test app copy
-          deploy = import_app(@options[:url] + 'deploy/app', body)
-          puts deploy.code
-          puts deploy.headers
-          puts deploy.body
-        end
+        sleep 1
+        importer
+        output.puts padder("Deployed #{@filey} to target APEX instance: #{@options[:url]}")
       end
     end
   end